+++ Patterns/PatternBuilderExternal.cs
@@ -15,7 +15,7 @@ namespace Patterns
             for (int i = letters.Length - 1; i >= 0; i--)
             {
                 int.TryParse(letters[i], out letter);
-                numSequenceBasic.InsertLetterMutable(0, (ulong)letter);
+                numSequenceBasic.InsertLetterMutable(0, letter);
             }
         }
