+++ Patterns/PermutationCheck.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
-using System.Diagnostics.Eventing.Reader;
-using System.Security.Policy;
+
 using NumericalSequences;
 
 namespace Patterns
@@ -12,12 +11,12 @@ namespace Patterns
             if (numSequenceBasic == null)
                 return false;
             
-            HashSet<byte> letters = new HashSet<byte>();
+            HashSet<int> letters = new HashSet<int>();
             int length = numSequenceBasic.Length;
 
             for (int i = 0; i < length; i++)
             {
-                byte letter = (byte)numSequenceBasic.GetLetter(i);
+                int letter = numSequenceBasic.GetLetter(i);
                 if (letters.Contains(letter) || letter >= length)
                     return false;
                 else
